requirements.txt
altgraph==0.17
ffmpeg==1.4
ffmpy==0.2.3
future==0.18.2
llvmlite==0.32.1
numba==0.49.1
numpy==1.18.4
pefile==2019.4.18
Pillow==7.1.2
PyInstaller==3.6
pywin32-ctypes==0.2.0
